Modified to hardcode the names of the pg branch (near the top of the file) and the webwork2 branch (in a commented out block, for use when the webwork2 tree should be pulled into the base image and not be bind mounted from the user's hard-disk.
These changes are roughly based on ones made by Michael Gage in
	https://github.com/openwebwork/webwork2/pull/877
but do NOT include the change to conf/defaults.config or the change
which added port 8080 to the WEBWORK_ROOT_URL setting, as those changes
did not seem to solve the issue with MathJax for me.

With the setting
	WEBWORK_ROOT_URL=http://localhost:8080
in the Dockerfile, searching using the library browser triggered
errors messages about the ajax queries. Dropping this change avoided
that issue.